Node v12 / Activation Xdebug à la volée

+++ build/conf/init.sh
@@ -5,7 +5,7 @@ ip route|awk '/default/ { print $3" host.docker.internal"}' >> /etc/hosts
 cp /var/www/conf/php/${PHP_CONFIG}/php.ini /etc/php/${PHP_VERSION}/apache2/conf.d/app.ini
 
 rm /etc/php/${PHP_VERSION}/apache2/conf.d/*-xdebug.ini
-if [ -f /var/www/conf/php/${PHP_CONFIG}/xdebug.ini ] && [ "${PHP_XDEBUG}" = "true" ]; then
+if [ -f /var/www/conf/php/${PHP_CONFIG}/xdebug.ini ] && [ "${PHP_XDEBUG}" = "true" ] && [ "${XDEBUG}" = "true" ]; then
   cp /var/www/conf/php/${PHP_CONFIG}/xdebug.ini /etc/php/${PHP_VERSION}/apache2/conf.d/20-xdebug.ini
 fi
